modified all 'delete' controllers in 'MedicalRecord'

+++ controllers/MedicalRecords/delete.js
@@ -5,6 +5,7 @@ const mongoose = require('mongoose');
 module.exports = async (req, res) => {
     const { _id } = req.body;
 
+    // check medical record's id
     if(!mongoose.Types.ObjectId.isValid(_id)) {
         return res.status(400).json({
             status: "error",
@@ -15,7 +16,17 @@ module.exports = async (req, res) => {
     }
 
     try {
-        const result = await db.deleteOne({ _id });
+        let result = await db.findOne({ _id });
+        if(!result) {
+            return res.status(404).json({
+                status: "error",
+                msg: `Data rekam medis tidak ditemukan`,
+                desc: null,
+                data: null,
+            });
+        }
+
+        result = await db.deleteOne({ _id });
         
         res.status(200).json({
             status: "success",
